% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p11-multiple-regression.R
\docType{data}
\name{d11.3}
\alias{d11.3}
\title{binary}
\format{A data frame with 400 rows and 4 variables}
\source{
\url{https://stats.idre.ucla.edu/stat/data/binary.csv}
}
\usage{
d11.3
}
\description{
\itemize{
 \item admit. 0 or 1.
 \item gre. gre score
 \item gpa. gpa score
 \item rank. 1,2,3,4, 1 is the highest, 4 is the lowest
}
}
\keyword{datasets}
